+++ src/runtime/composables/useChatgpt.ts
@@ -1,14 +1,18 @@
 import { createError } from 'h3'
-import type { IChatgptClient, IMessage } from "../types"
+import type { IChatgptClient, IMessage, IModel, IOptions } from "../types"
 
 export const useChatgpt = (): IChatgptClient => {
 
-  const send = async (message: IMessage) => {
+  const chat = async (message: IMessage, model?: IModel, options?: IOptions) => {
 
     try {
-      return await $fetch('/api/openai', {
+      return await $fetch('/api/chat', {
         method: 'POST',
-        body: message
+        body: {
+          message,
+          model,
+          options
+        }
       })
     } catch (error) {
       throw createError({
@@ -18,5 +22,24 @@ export const useChatgpt = (): IChatgptClient => {
     }
   }
 
-  return { send }
+  const chatCompletion = async (message: IMessage, model?: IModel, options?: IOptions) => {
+
+    try {
+      return await $fetch('/api/chat-completion', {
+        method: 'POST',
+        body: {
+          message,
+          model,
+          options
+        }
+      })
+    } catch (error) {
+      throw createError({
+        statusCode: 500,
+        message: 'Failed to forward request to server',
+      })
+    }
+  }
+
+  return { chat, chatCompletion }
 }