+++ src/runtime/constants/models.ts
@@ -1,0 +1,11 @@
+// chat
+export const MODEL_TEXT_DAVINCI_003: string = 'text-davinci-003'
+export const MODEL_TEXT_DAVINCI_002: string = 'text-davinci-002'
+
+// chat completion
+export const MODEL_GPT_TURBO_3_5: string = 'gpt-3.5-turbo'
+export const MODEL_GPT_TURBO_3_5_0301: string = 'gpt-3.5-turbo-0301'
+export const MODEL_GPT_4: string = 'gpt-4'
+export const MODEL_GPT_4_0314: string = 'gpt-4-0314'
+export const MODEL_GPT_4_32k: string = 'gpt-4-32k'
+export const MODEL_GPT_4_32k_0314: string = 'gpt-4-32k-0314'