+++ src/runtime/types/index.d.ts
@@ -1,8 +1,20 @@
 export interface IChatgptClient {
-  send ( IMessage ) : Promise
+  chat(IMessage): Promise,
+  chatCompletion(IMessage): Promise
 }
 
 export interface IMessage {
   message: string
 }
 
+export interface IModel {
+  model: string
+}
+
+export interface IOptions {
+  temperature: number,
+  max_tokens: number,
+  top_p: number,
+  frequency_penalty: number,
+  presence_penalty: number,
+}